style. common - Category component
category 페이지 및 category 페이지 내 CategoryBar, CategoryContents 컴포넌트 네이밍과의 충돌을 피하기 위해 컴포넌트명 수정
기존 common/Category
수정 common/CategoryBox

Related to : #99
